Correções Feature crud transacao pix

+++ src/Cepedi.BancoCentral.PagamentoPix.Compartilhado/Responses/CriarTransacaoPixResponse.cs
@@ -1,2 +1,2 @@
 ﻿namespace Cepedi.BancoCentral.PagamentoPix.Compartilhado.Responses;
-public record CriarTransacaoPixResponse(int idTransacaoPix, double valor);
+public record CriarTransacaoPixResponse(int idTransacaoPix, decimal valor);
